rustlib: impl. Object#==, Object#puts
