+++ lib/skc_rustlib/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rustlib"
+name = "skc_rustlib"
 version = "0.1.0"
 authors = ["Yutaka HARA <yutaka.hara+github@gmail.com>"]
 edition = "2018"
