+++ lib/skc_rustlib/src/builtin.rs
@@ -1,0 +1,5 @@
+pub mod bool;
+pub mod int;
+pub mod object;
+pub mod ptr;
+pub mod string;
